make ruff and mypy pass again after a toolchain update
Both targets started failing on a clean tree once newer ruff and numpy
landed in the dev environment.  Neither failure was new code; both were
latent.

tests/test_wfmconvert.py put a nested f-string with line breaks inside a
replacement field, which PEP 701 allows only from Python 3.12.  This
package declares requires-python >= 3.10, so the test suite could not
even be parsed on the two oldest versions it claims to support.  Newer
ruff, which infers its target from requires-python, is right to reject
it.  Hoisting the inner string out of the field fixes it, and a scan of
every module confirms nothing else uses syntax newer than the floor.

numpy 2.5 writes its bundled stubs with PEP 695 `type` statements, which
mypy refuses to parse while assuming python_version 3.10, so it bailed
out before checking anything.  mypy now assumes 3.12; ruff still holds
the package to the 3.10 syntax floor, so that side is still covered.

Letting mypy past the stubs surfaced a real error it had never reached:
mso5074.py stored float64 volts in channel_data, which is declared as
float32 and is float32 in every other adapter.  Its samples are 8-bit
ADC counts, so the cast costs nothing.

+++ RigolWFM/mso5074.py
@@ -145,7 +145,9 @@ def from_file(file_name: str) -> Mso5000Waveform:
 
         # Convert raw uint8 ADC counts to approximate volts.
         # No calibration coefficients are present in the file.
-        volts = (samples.astype(np.float64) - _ADC_MIDPOINT) / _COUNTS_PER_VOLT
+        # float32 to match `channel_data`, as every other adapter does; the
+        # source is 8-bit ADC counts, so nothing is lost.
+        volts = ((samples.astype(np.float64) - _ADC_MIDPOINT) / _COUNTS_PER_VOLT).astype(np.float32)
 
         ch_name = str(wh_fields["waveform_label"]) or f"CH{slot + 1}"
         channel = ChannelHeader(ch_name, enabled=True, unit_code=1)  # 1 = V
